+++ encoding/protobuf/internal/testpb/v2/test.pb.yarpc.go
@@ -25,7 +25,7 @@ package testpb
 
 import (
 	"context"
-	"io/ioutil"
+	"io"
 	"reflect"
 
 	"go.uber.org/fx"
@@ -37,7 +37,7 @@ import (
 	"google.golang.org/protobuf/proto"
 )
 
-var _ = ioutil.NopCloser
+var _ = io.NopCloser
 
 // TestYARPCClient is the YARPC client-side interface for the Test service.
 type TestYARPCClient interface {