+++ encoding/protobuf/protoc-gen-yarpc-go/internal/lib/lib.go
@@ -47,7 +47,7 @@ import (
 	{{range $i := .Imports}}{{if not $i.Standard}}{{$i | printf "%s\n"}}{{end}}{{end}}
 ){{end}}
 
-{{if ne (len .Services) 0}}var _ = ioutil.NopCloser{{end}}
+{{if ne (len .Services) 0}}var _ = io.NopCloser{{end}}
 
 {{range $service := .Services}}
 // {{$service.GetName}}YARPCClient is the YARPC client-side interface for the {{$service.GetName}} service.
@@ -625,7 +625,7 @@ var Runner = protoplugin.NewRunner(
 	checkTemplateInfo,
 	[]string{
 		"context",
-		"io/ioutil",
+		"io",
 		"reflect",
 		"github.com/gogo/protobuf/jsonpb",
 		"github.com/gogo/protobuf/proto",